Fixed average fitness bug

+++ src/genoperators.h
@@ -3,7 +3,7 @@
 	
 void mutate(int *,int ,double );
 void crossover(int *,int *,int );
-void roullete_wheel_selection(int **,int ,int ,int **,int *);
+void roullete_wheel_selection(int **,int ,int ,int **,double *);
 void tournament_selection(int **,int *,int  ,int ,double );
